refactor operator

+++ src/operators/tensor/core.cairo
@@ -4230,7 +4230,7 @@ trait TensorTrait<T> {
     /// # tensor.gather_elements
     ///
     /// ```rust 
-    ///    fn gather_elements(self: @Tensor<T>, indices: Tensor<T>, axis: Option<usize>) -> Tensor<T>;
+    ///    fn gather_elements(self: @Tensor<T>, indices: Tensor<i32>, axis: Option<i32>) -> Tensor<T>;
     /// ```
     ///
     /// GatherElements is an indexing operation that produces its output by indexing into the input data tensor at index positions determined by elements of the indices tensor.
@@ -4275,7 +4275,7 @@ trait TensorTrait<T> {
     ///      [7. 2. 3.]]
     /// ```
     ///
-    fn gather_elements(self: @Tensor<T>, indices: Tensor<usize>, axis: Option<usize>) -> Tensor<T>;
+    fn gather_elements(self: @Tensor<T>, indices: Tensor<i32>, axis: Option<i32>) -> Tensor<T>;
     /// # tensor.binarizer
     /// 
     /// ```rust
