+++ src/operators/tensor/implementations/tensor_fp8x23wide.cairo
@@ -393,7 +393,7 @@ impl FP8x23WTensor of TensorTrait<FP8x23W> {
     }
 
     fn gather_elements(
-        self: @Tensor<FP8x23W>, indices: Tensor<usize>, axis: Option<usize>
+        self: @Tensor<FP8x23W>, indices: Tensor<i32>, axis: Option<i32>
     ) -> Tensor<FP8x23W> {
         math::gather_elements::gather_elements(self, indices, axis)
     }
